refactor: useNavLinks custom hooks
