+++ package.json
@@ -12,6 +12,7 @@
     "react-scripts": "5.0.1",
     "styled-components": "^5.3.5",
     "styled-reset": "^4.3.4",
+    "uuid": "^8.3.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -53,6 +54,10 @@
             [
               "context",
               "./src/context/"
+            ],
+            [
+              "hooks",
+              "./src/hooks/"
             ]
           ],
           "extensions": [
